Reorganise the mass of supporting benchmark functions
